backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
openai>=1.30.0
faiss-cpu
numpy